R: split magic command tests into separate files
